+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the Zowe CLI package will be documented in this file.
 
+## Recent Changes
+
+- Added job statement to the workflow creation command (`zowe wf create`) as `workflow-job-statement`
+
 ## `7.0.0-next.202012221518`
 
 - Enhancement: Added daemon mode proof of concept.